python/requirements.txt
conductor-python==v1.0.7

opencv-python
pytesseract
